Problem: pg_yregress test group failure can be verbose
When a sub-test fails in a group, it gets its nice expected/result YAML
report to compare. However, its parent test will also fail (which is
fine) but it'll print a much larger YAML report which obscures the
smaller output and makes it harder to discover the actual failure.

Proposed solution: ensure parent tests don't print the expected/result
YAML report on failure as sub-tests already do that.

It isn't perfect yet, I think, but seems to generally do the job. Area
worth exploring is what happens if the `step` is also `commit: true` and
it fails during the commit.

Fixes #222