Problem: omni_vfs method dispatch
omni_vfs uses a technique where we dispatch all FS-specific calls to
omni_vfs.method. The problem is that since some of these functions may
use types defined in the same schema by the omni_vfs extension, one
can't drop the extension without dropping these functions altogether,
which may not be a great DX in the sense that if we had to drop/create
omni_vfs, we would have to re-create these functions.

Solution: extract omni_vfs types into omni_vfs_api extension

This is not ideal (one more extension) but it does accomplish the job
to a degree.

Addresses #219