Problem: updating omni_httpd handler for a listener on an effective port
If omni_httpd listener has a port specified (say, 8080) but it wasn't
available to listen on, an effective port will be picked using operating
system's facilities.

However, a subsequent update of a handler triggers omni_httpd to try
listening 8080 again (which is not a bad idea to try and return it back
to where it is intended to be), but if that fails again, a new port will
be picked, which is suboptimal and (frankly) annoying.

Solution: try seeing if we're still listening on the effective port

If we do, use it. If we don't try to listen on it. If that doesn't work,
let the operating system choose the port.

NB: there is currently no test for this behavior because it is
predicated on the test to have a certain port listened on in any test
environment (so that it returns `EADDRINUSE`) and I don't (yet?) see a
good way to do this.