Problem: omni_vfs.file_info errors on file not found
This is difficult to deal with when writing a static HTTP handler
because it requires exception handling which is only available in
PL/pgSQL or other languages, but not SQL.

Solution: make it return NULL instead

May be not a perfect solution but makes sense for now. May consider a
more sophisticated type for it later.