Problem: debugging errors in pg_yregress
When we're writing a test and it fails, the resulting output is not very
useful. We get `success` set to `false` but it doesn't really give us
much. This requires one to go back to the test definition, add an empty
`error` key and re-run the test to get the actual error. After figuring
the issue out, this change needs to be reverted

This is hardly practical and is annoying. Not to mention that it is
confusing that the single concept of failing is managed by two different
keys.

Solution: drop `success` altogether

This is achieved by allowing `error` to be set as a boolean if details
of the error are not of interest. Thus, one can set `error` to `true`
instead of `success` being set to `false`. When `error` is not set, it
will be populated with the scalar representation of the error.

During this implementation, two important bugs were fixed:

* `negative` flag was never set correctly
* matching boolean scalars was failing to do a correct job to ensure
  full equality of strings, thus "f" was always matching "false" and so
  on, and that was unintentional.