Problem: subtests with a failing scalar test loop indefnitely
This happens if we have something like this:

```yaml
- tests:
  - slect
```

Solution: ensure to mitigate lack of node replacement

Currently, because of lack of node replacement
(https://github.com/pantoniou/libfyaml/issues/84), when scalar nodes
are replaced with mapping (containing the error), they are actually
removed and a new node is inserted. However, this came into conflict
with subtest iteration.