Problem: serving static files with omni_httpd
This requires writing the server yourself. People shouldn't be doing
this.

Solution: provide a default implementation

Can be used like this:

```postgresql
update omni_httpd.handlers
    set
        query = (select
                     omni_httpd.cascading_query(name, query order by priority desc nulls last)
                 from (select * from omni_httpd.static_file_handlers(mount_point(), 0)
                       union (values
                                  ('test',
                                   $$ select omni_httpd.http_response('passed') from request where request.path = '/test'$$, 1))) routes);
```

I am not 100% happy with putting `static_file_handlers` into omni_httpd,
but until I can figure out a better place for it, I guess that's where
it is going to be.

One thing that bothers me more here is how we pass the `fs` value
around: we pass it as a string literal and cast it to the type of it.
Frankly, this feels fragile.