Problem: passing fs to static_file_handlers as is
If we ever need to change the underlying file system, we'd need to
change it in every place where we passed it to.

Solution: pass a regproc instead (as a "getter")
so that we can change the return value to do so.

This is still not ideal as we check regproc by OID but then refer to it
by name (as the OID may change), but it is better than before, or so it
seems.