Problem: no support for Postgres 16
We really want 16 as it has a lot of significant improvements.

Solution: include it in the build

The most challenging piece here is updating omni_sql deparsing. It's not
perfect yet and doesn't support all the changes but it should serve well
for the initial goal of supporting Postgres 16.

Docker container temporarily uses our own build of postgres 16 container
until Docker Hub gets 16.0 in.

I removed pljava for the build (temporarily) as the distribution doesn't
have it yet.