Problem: typing Python classes with composites
If we want to type a Python class and use it as a composite, we have to
type hint it as `Composite(<type>, "name")` which is diverting us from
writing a simple time. Even if we assign `Composite(...)` to something,
it's still a different name than the original type.

Solution: classes can provide optional type hint

This is done by adding a class method of `__pg_type_hint__` that must
return a type signature.

Currently, the extractor is still limited and imports done in the module
don't seem to be working inside of this class method. Local imports
work. For this reason it may be worth considering making this hint an
annotation-only hint without referencing the class itself. The jury is
still out on what's the best method here.