Problem: omni_httpd producing duplicate `content-length`
If HTTP handler has produced `content-length` header, omni_httpd will
add its own based on the size of the response. This creates duplicate
headers.

Solution: check for `content-length` in the response

If the size specified is smaller, cut the response body to fit into it
to avoid overflow. If the specified size is too large, send a warning
but let omni_httpd calculate the right size.

In the end, continue letting h2o handle the header.