Problem: impossible to grow sum_type
However, sometimes this is very desirable as types evolve over time.

They can't grow because if they are not variable length, their length is
limited to the length of the type tag and the length of the largest
variant in storage.

Solution: provide a renaming workflow

* `omni_types.rename(regtype, name)` will rename the type
* the tables using the old type can be then altered:

```sql
alter table tab alter value type type1 using (value::text::type1)
```