Problem: omni_httpd fallback strategy is flawed
It will attempt to fall back to effective port even if it is the same
port as we just tried to listen at and failed.

Solution: ensure these ports aren't the same

If they are the same, try picking a new port.

(I wonder if we can come up with a good way to test behavior of this?)