Problem:  errors during omni_httpd restart post-crash
It seemingly attempts to bind the port multiple times and is having
difficulties with that.

Solution: ensure it is only omni_ext that manages the lifecycle of omni_httpd
and other Dynpgext-registered background workers

Importantly, we do the same for omni_httpd-controlled http_workers as
well, otherwise, it exhibits the same problem which manifests in
postgres having trouble terminating cleanly afterwards.