Problem: omni_python configs don't survive the restart
That's because they are set using `set_config`

Solution: persist configuration in omni_python.config

Also, make site_packages default to `data_dir/.omni_python/default`
instead of the user's home directory. This way we're not polluting where
we shouldn't and when used with Docker, it'll survive in the data volume.