Problem: HTTP endpoint to serve local Python packages
For pre-release scenarios we need to be able to use locally built
versions of Python packages. This currently requires running an HTTP
endpoint to serve this, and this is a little bit excessive and
error-prone.

Solution: use pip's `--find-links` option

This allows us to point to a directory with wheel files. This is not
ideal (can we avoid doing anything on the filesystem?) but this is way
better than having to have an HTTP endpoint just for this.