Problem: do we create correct S3 PutObject requests?
Our tests don't try to upload actual (non-textual) binaries. Does the
signing work correctly there? We don't have a clear answer to this.

I have a report that at least such requests resulted in a signing error
and trying to investigate.

Solution: write a test and ensure it works in a basic case