Problem: top-level Makefile confuses people
Some people attempt to use it without knowing its purpose (not that it
is explicitlyi indicated anywhere) and seeing errors like #339

Solution: get rid of it for the time being

It has been previously used to test Omnigres against different versions
of Postgres, but I haven't used it much lately and I assume nobody else
did. GitHub Actions CI is a (slower) alternative to this for the time
being.