Problem: Python code importing local modules
omni_schema + omni_python do not collaborate to provide local imports.
One has to resort to providing path-based dependencies through
requirements.txt.

Solution: implement a custom importer

The importer requires a (currently janky [1]) integration with omni_vfs.

The importer would only consider files in the same directory, no
submodules or directories. This is just to get going. This has to be
fixed.

[1] We can't pass `anyelement` in plpython so we're resorting to pass
the textual representation of the fs and its type and then sticking that
into a new query. Would having our own fork of plpython resolve this?