Problem: wrapping callables in Python is a bit annoying
Currently, if we want to expose a callable to Postgres, we have to do
something like this:

```python
thing_ = Thing()

def thing(in: int) -> int:
  return thing_(in)
```

This is particularly common for Flask integration.

While certainly doable, it is a bit annoying/redundant.

Solution: make any callable a fair game

Now it is possible to wring the code above as:

```python
thing = pg(Thing())
```