Problem: importing request from flask
At the time of the loading a Python file with such import, we'd get the
following error:

```
RuntimeError: Working outside of request context.

This typically means that you attempted to use functionality that needed
an active HTTP request. Consult the documentation on testing for
information about how to avoid this problem.
```

Solution: don't consider callables without any attributes

If a callable has an empty `dir`, then there's no `__pg_stored_procedure__ and we can avoid
calling `hasattr` which triggers the above error.