Problem: Python files with local imports
If one uses omni_schema to provision python code and it uses modules
defined in the VFS, such code will not survive backend restarts.

This means that event simply [re]connecting to the database renders
these function calls broken. So does an attempt to use them in
background workers (say, omni_httpd).

This is pretty critical.

Solution: ensure we keep the dependencies around

All the modules that were imported through VFS (using custom importer)
will be cached in the target script so that a new a backend has
something to pick up.

This is still not ideal as it doesn't handle imports that happen during
runtime (as opposed to script evaluation time) but it's better than
nothing as it covers majority of cases.