Problem: reports of omni_ext/httpd crashing postgres
This happens if Postgres is built with assertions enabled.

Solution: enable assertions and figure out the crashes

1. Ensure we always set hash element type (there are assertions checking
   for defaults)
2. Ensure we allocate background worker handles in top memory context to
   avoid reclaiming them too early.
3. Ensure we restore user and security context after committing
   transactions (abort reverts them, commit doesn't)

Also, discovered a case when omni_httpc client is used with null queries
only: it'll crash. Avoid doing anything on requests if there are no
actual requests to make.

Importantly, this makes Postgres builds slower and these are not as
suitable for benchmarking as before. However, correctness being more
important, this is a better default for development. If you need to
benchmark, build with `-DCMAKE_BUILD_TYPE=Release` or `RelWithDebInfo`.