Problem: omni_aws S3 doesn't support DigitalOcean Spaces
This is because it uses subdomain-style bucket naming and we only do
support path-style for anything that is not AWS S3.

Solution: make endpoint templates with region and bucket placeholders

This changes `endpoint_url` parameter to `endpoint`.

While at it, we needed to unparse URLs, so omni_web was extended to
support this case.