Problem: finding the correct version of OpenSSL
This is often a problem as target systems may have a wrong version, and
discovery methods vary between systems and installation methods.

Solution: bring OpenSSL build to the configure stage

This also builds Perl to avoid depending on system Perl and
Text::Template. On its own this seems a bit ridiculous, but considering
we will be building PL/Perl, this would make more sense in the future.

This increases the configure time but (hopefully) makes builds more
reliable.