Problem: omni_var crashes in Postgres <= 15
This happens when `set()` was used when the transaction is committed or
aborted.

Solution: avoid de-registering the callback in these versions

As of
[4d2a844242dcfb34e05dd0d880b1a283a514b16b](https://github.com/postgres/postgres/commit/4d2a844242dcfb34e05dd0d880b1a283a514b16b) in Postgres
it is possible to unregister a callback in itself. Prior to that, it's
not.

As of now, we'll always keep this callback registered as there's no way
it can be unregistered from itself. It's theoretically possible to solve
this by adding a Postgres pre-16 function that would unregister the
callback but it seems like asking for trouble.