Problem: omni_var may be copying memory unnecessarily
Currently, transactional variables copy the values when a new
sub-transaction is started, and the value is overwritten (or when we
roll back to a savepoint (see uses of `memcpy`).

While it is not a significant concern, it is still less efficient than
one would want. The size of `VariableValue` on a 64-bit platform could
be 32 bytes.

Solution: always allocate `VariableValue`

This forces the variable value to always go through a pointer to get to
the value. The initial allocation comes from the allocation done for
`Variable`.

This way, we don't need to copy the variable value and simply change
pointers.