Problem: listing extensions in CMakeLists.txt manually
This is just a bit annoying, especially when working on new extensions
in long-living branches and having to resolve conflicts in
extensions/CMakeLists.txt

Solution: find extensions automatically

This also highlighted the fact that the dynpgext test didn't depend on
omni_ext (and it should have). This has been fixed now.