Problem: omni_vfs backend support is still limited
Currently, omni_vfs only has only one backend: local_fs.
This is very limiting as this is not as practical on the server.

Solution: implement a table_fs that will provide the same API for table-backed filesystems

path parameters accepted by the functions is converted to
absolute canonical form before usage

Signed-off-by: Rahul Modpur <rmodpur2@gmail.com>