Problem: preloading some extensions in tests
If an extension is not a dynpgext extension, we need to be able to load
it without omni_ext. But we don't know the filename because it contains
a version.

Solution: pass `EXTENSION_FILE` environment variable to pg_yregress

It can now be used as `*/env/EXTENSION_FILE` in the test