Problem: manually trying certain behaviours between restarts
One problem with `make psql_<EXT>` workflow is that it re-initializes a
database every time it start. This makes it hard to explore scenarios that
require restarting the database and still having the same data.

Solution: introduce PSQLDB environment variable

Specifying this variable instructs the aforementioned `psql` targets to
use a specific location for the database and don't wipe it.