Problem: upgrading between arbitrary versions of extensions
It's problematic as there are no upgrade scripts.

Solution: generate upgrade scripts "on demand"

The general idea here is that we can compare the installations of each
version and find the source code of which functions ere changed. We can
also see which part of the incremental migration was added.

What we haven't considered here yet is what happens if the version
introduces additional function signatures to the schema. Ideally, they
shouldn't do this as part of `src` includes but in a new incremental
migration.

This implies that we no longer permit arbitrary changes in incremental
migrations between releases. They have to be incremental. The only thing
that stays away from this is the code of the functions.