Problem: loading multiple versions of omni_httpd
When doing so, the started version tends to get stuck after "Established
listener on PORT" and not serve anything.

Solution: ensure we always determine the number of workers

This is now accomplished by looking at the string value. This is
suboptimal as we're not able to observe changes to this the same way a
GUC observation would, but it's still doable.

More context in #447.