Problem: upgrading omni_httpd in runtime
There is no way to stop it to load another version of the extension.

Solution: use omni_ext unloading to kill the master worker

The master worker's PID is set by the master worker itself upon startup.

Also fixes a bug in omni_ext that wouldn't really unload anything that
was already loaded during startup.

This is not a complete end-to-end solution to upgrading omni_httpd, but
it is definitely a component to it.