Problem: installing or upgrading multiple extensions
It is troublesome to trace the requirements of all extensions to install
or upgrade them correctly.

Postgres doesn't really support requirement versioning and cascading
installations are only helpful to a degree (not when versions come into
the picture)

Solution: provide a way to install/upgrade a manifest

A manifest contains a list of all requirements to install along with
their requirements. Think of a dependencies lock file.

This is implemented as a PL/pgSQL extension to allow using it in any
installation of Postgres without having to depend on native code or
additional language support.