Problem: Scheduled Builds require approval
Context: It looks like scheduled builds were introduced to keep the GitHub build cache warm. Later an approval step was introduced to the workflow which requires approval even for the scheduled builds, which results in those builds not being run without approval.

Solution: Run scheduled builds on the `master` environment. 

Alternative Solution: Alternatively we can delete the scheduled cron because this particular workflow has been migrated to Warp and FlyCI.