Problem: omni_ext does not start database-local workers in runtime
When an extension is not preloaded, database-local workers are not
started since we need to scan extensions in different databases.

Solution: use a shared table to list background worker requests

This also avoids the need for a rather convoluted rendezvous we had.

I got rid of callbacks in background worker registration. While this has
worked for workers initialized during the initial startup phase, this
has become increasingly more complex for runtime-started workers.
Problem is that we can't simply point to the function or the callback
payload if they weren't loaded and prepared during the startup phase in
the parent process. To date, this capability hasn't been used and, as
far as I can tell, it doesn't provide any extra functionality that can't
be achieved by executing that logic in the worker itself.

Closes #21
Closes #22