Problem: omni_ext background worker startup not tested correctly
Specifically, the test for global background worker for preloaded
extensions pre-creates a table in its schema and the table is used as
a marker. So it never really tests it.

Also, to add, recent change that passed db_oid to global worker won't
work during preloading as `MyDatabaseId` at that time is 0 (if it is
actually initialized at the time we prepare the background worker entry)

Solution: introduce some "reserved" space in SharedInfo of omni_ext

This is a strictly internal hack for the time being that allows tests to
have a shared location in memory to write something to.

The test will now write a flag there once the global background worker
is started. A new utility function has been added to track that.

This could have been done better, but I am a bit tired and I need this
to work right now.