Problem: omni_ext error of "Not enough workers allowed"
It happens when there are a lot of extensions being preloaded. However,
we are not attempting to load all of them (which would, naturally,
result in the exhaustion of worker slots if they register all of their
background workers.)

Solution: drop the check

Pre-loading only means we've pre-allocated some resources to them but
doesn't mean we'll ever register all the workers.