Problem: installing a set of Omnigres extensions
It's still difficult to just take a version of Omnigres and install or
upgrade a set of extensions from it.

Solution: provide exchange formats for artifact listing

Manifests are effectively artifact lists, so we now support text and
JSON formats for describing artifacts. Omnigres build process produces
such artifact lists upon each build (`artifacts.txt`).

The delivery of this file may depend on a particular process as we don't
have a common one established yet. One can put it onto the file system
and read it with `pg_read_file`, or can download it using `omni_httpc`
(if they have it already, that is).

Quite a lot going on in this change besides this:

* It fixes an issue with omni_ext hooks incorrectly handling the case of
  an attempt to upgrade version of an extension to the same version,
  assuming there's an unloading to be made where there shouldn't be.
  The executor as it is doesn't fail to prevent this but just issues a
  notice.
* New requirement status `kept` added to signify the above case