Problem: occassional FATAL error in omni_httpd's master worker
The error is: "cannot read pg_class without having selected a database"

Solution: ensure we start the worker with the correct database OID

Turned out that in some cases, we were going through the
`register_bgworker_runtime` path and it was not registering the intended
description of the background worker (`local_bgw`) but the original one
that didn't have a correct main argument set.

With this change, it bow registers the correct one.

This applies to other dynpgext components beyond omni_httpd.