Problem: omni_ext has a number of flaws
While it does it job (with some known quirks and bugs), it has grown to
be a little to clunky for our evolving needs.

For example, it's scheme of "preloading" extensions is problematic: it
preloads extensions that are not installed (wasting resources and
hurting the load time), it does not solve the problem of centralized
hook injection, and so on.

Its core was developed about a year ago, and we've learned how to build
complex extensions better.

Solution: introduce a new, refined extension

Say hello to `omni` extension and interface. Upon maturity, they are
replacing `omni_ext` and `dynpgext`.

Seeing how `omni_ext` has recently benefited from hooks, my initial
focus is on providing hooking capabilities for extensions. While I
haven't implemented every hook in the book in this commit, it shows the
general direction and more hooks will be implemented down the road.

This is still in an early phase and it doesn't have all the capabilities
of `omni_ext` (shared memory, bgworkers) but it is showing promise
already. As you can see, there are still some parts of omni.h that were
inherited from dynpgext.h and they are going to evolve as we transition
and refine.

`omni_ext` and `dynpgext` aren't deprecated yet as other components
depend on them.