Problem: omni is initializing modules not enabled in a backend
Consider the following situation:

1. We have a database D1. We provision an extension that loads a module
   M.
2. We now create database D2 and connect to it.
3. It now initializes M for this backend!

This shouldn't really happen as this backend is not yet enabled in this
database.

Solution: cross-reference the necessity of loading a module
when scanning for newly available modules

The implementation is not perfect as it rescans `pg_proc` for `probin`
to see if anythign is requiring this module. But this is something we
can improve going forward, I suppose.