Problem: `omni` does not provide shared memory allocation primitives
In order for it to be able to replace `omni_ext`, it should provide such
functionality as it is very useful.

Solution: review and redesign the approach in `omni_ext`

Most notable difference is that (besides moving the API into the handle)
we no longer provide scoping (database/local). This is because it would
make the code a lot more complicated, while the same can be achieved by
regulating the name of the shared memory allocation.

Internally, the implemented has changed from using a pre-allocated pool
of memory to DSA.