Problem: WaitForBackgroundWorkerStartup hangs
This happens when it is started from `database_worker`, i.e. at the very
startup of the cluster.

Solution: ensure we don't deadlock and unblock signals

This also fixes a rather grave mistake of forgetting that the locks are
returned padded and thus doing a simple offset from one lock to another
won't work as it won't point to the next lock. This has also caused
lockups.