Problem: BackgroundWorkerHandle is a private type
It means we can't easily share it between omni-powered backends as we
don't know its size.

It means we can't really allow backends to use
BackgroundWorkerHandle-centered API on background worker, which is an
unpleasant limitation.

Solution: track the layout of BackgroundWorkerHandle across versions
and introduce header-only API in omni.h to circumvent the limitation.