Problem: handling loading/unloading of omni modules
Currently, there are multiple problematic cases I can observe:

* we first unload the module (`_Omni_unload`) and never deinitialize (by calling
  `_Omni_deinit`)
* when some backend unloads a module, hooks get removed in other
  backends

Solution: track the state of whether a module is loaded

This allows us to unload module once, but we don't guarantee the order
in relation to deinitialization. The main purpose of this callback is
that it is called once.

I've further sped up and simplified the refreshing of modules by using
syscache callback to avoid rescanning `pg_proc` ever so often.

In order to make use of `_Omni_deinit` practical, I've introduced shared
memory deallocation functionality, so that `deinit` can free up memory
and also observe the leadership in that resource deallocation to take
action at most once (for example, shutting down a background worker.)