Problem: declaring and using GUCs from extensions
This is quite painful. While doable, one needs to:

1. Use different functions for registering variables of different types
   (with multitudes of parameters)
2. Detect if a config has been previously defined and if yes, coordinate
   use of the pointer to update value at

Solution: introduce unified GUC API in `omni`

This gives a single point of entry access to declaration and retrieval
of GUC pointers, regardless of whether the variable was declared before
or not. Omni handles the allocation of memory for these GUCs.

The API uses a type with enums to conceal "default" parameters and
indicate configuration. This exposes the total of one function.

This is not yet well documented. Soon.