Problem: testing gh workflows before merging to master is really difficult
One has to merge their workflow changes to master to see it run
which leads to trial and error approach to get it working correctly.

Solution: add a dummy branch target "test-gh-workflows" for push events
to test workflow changes before merging them to master. Only adding it
to Build workflow for now, we can always add it to other workflows later
if required.

Signed-off-by: Rahul Modpur <rmodpur2@gmail.com>