Problem: omni_httpd still using `omni_ext`
We're transitioning to `omni` and `omni_ext` will not receive any
updates.

Solution: switch it (and omni_containers) to `omni`

There are some pieces in `omni` that make it less than ideal as this
transition shows. We'll address them as follow-ups.

`omni_containers` technically dependended on `omni_ext` but didn't use
it.

Removed `omni_ext` and `dynpgext` as they are no longer used.

Fixed an issue with potential deadlock on access exclusive.