Problem: using omni.h as the include for omni interface
Omni versions will change and requiring the same file can get very
confusing with regard to adherence to a particular version of the
interface.

Solution: move omni.h to omni/omni_vRELEASE.h

Where release is either zero (0), a special "placeholder" version for
the release until we cut the first release of the interface, or a
combination of a number and a letter:

1A, 1B, 1C, 2A...

Number signifies breaking backward compatibility while the letter
signifies the revision within this compatibility band. Only incremental,
backward-compatible changes can happen there.