Problem: use of fixed-size hash tables to store module handles
Omni stores module handles and module hash tables in fixed-size shared
memory segments. This introduces a physical limit to have many modules
can be used. It is set pretty high, so it is hard to hit this number,
but this also wastes memory when it is underutilized.

Solution: use dshash and DSA instead

This way, we're allocating memory segments as necessary. The code is
tiny bit more complex (particularly around dshash release of entries),
but allows for such granularity.

P.S. Had to include dshash.c from Postgres to get support for
dshash iteration for Postgres older than 15. Any better ideas?