Problem: intermittent omni_httpd master worker crashes
Happens when it tries to read GUC for `temp_dir`

Solution: ensure it only does so if the extension is preloaded

When register the background worker, it is not started immediately but
we rather just ask the postmaster to start it and then we proceed
without waiting.

So, if we drop the extension shortly after, it is not going to be seen
by the background worker and it should just terminate.