Problem: omni hook name becomes garbage
You can see that in `omni.hooks`

Solution: introduce module's own memory context
and run main callbacks in it.

Hook registration now requires `hook->name` to outlive the lifecycle of
the module in the backend. This can be done by allocating within the
module's own memory context or a longer-living context (such as
TopMemoryContext) or static allocation.

The introduction of module's own memory context opens a nice way for the
module to allocate for the lifetime of the module and get cleaned up
upon deinitialized without any extra work.