Problem: upgrading extensions without touching functions
When Postgres upgrades an extension to a version that has a different
`module_pathname`, it doesn't touch functions defined by the old
version. That is, all records in `pg_proc` remain pointing to the shared
object file from the old version, which is highly erroneous and can lead
to undefined behavior.

The reason why we choose different shared object file also has to do
with correctness, stability and predictability. Replacing a single .so
file changes the implementation for a running cluster (which may
potentially want to have multiple versions of the extension installed)

Solution: update relevant pg_proc entries on upgrade

We're doing this by hooking into `alter extension` utility statement and
traversing `pg_depend` for function that depend on the extension and
changing their `probin` column to reflect what's in the new
`module_pathname`.