Problem: resource cleanup in omni is still problematic
For example, what happens when we need to deallocate shared memory?

We can remove it from the allocation table by a single backend, but then
how do other backends know that allocation is gone without consulting
the table? This may very well lead to a crash.

Solution: reference count memory deallocation and introduce atomic switches

We will now deallocate memory only when all backends that use the module
have deallocated it. We will still delete the record from the table on
the first attempt, but we will not physically delete it until
everybody's good to go.

Importantly, to control the state of the module, we now have an atomic
switchboard. This allowed us to get rid of `_Omni_load` and
`_Omni_unload` callbacks as the behavior exposed by these is now
trivially emulated with atomic switches.