Problem: omni init/deinit callback order
Due to the way omniis implemented, we will first call `init` on newly
discovered modules and then `deinit` on removed modules.

However, this is not going to work well for extensions that rely on
previous versions to be removed first, if they are competing for a
shared resource. A great example of this is omni_httpd. For it to work
properly, the listener need to shut down first before attempting to
listen again.

Solution: move initialization past deinitialization

On the surface, this change is trivial (instead of calling `_Omni_init`)
when we discover the module, call it once the entire list is available,
after deinitialization.

However, there were a few issues to resolve here.

Firstly, when we update probin upon upgrade, the changes were not
visible to the pg_proc scanner. Incrementing command counter on update
helped here.

Secondly, I discovered that some of our hooks were not correctly
attempting to induce rescanning. Specifically, when I added SPI usage to
init/deinit in the test (to track the order of operations), Postgres'
assertions started to fail. Turns out, we were rescanning during
`TransactionStmt` processing and this, to the best of my understanding,
was leading to running transactional logic in an environment where not
all invariants are observed. We no longer try to rescan during these
statements.

I also had to deal with the order of registration for worker termination
and startup. The problem was obscured by the fact that when we register
a xact callback, they get prepended to the list (so they are executed in
the reverse order), thus leading to potential startup of new workers
before shutting down the old ones (that is critical to omni_httpd).
However, in the specific case of omni_httpd this was mitigated by the
fact that it is requesting operations to happen _after the commit_,
which involves the use of the TopTransactionContext slingshot. The
callbacks registered there are executed in the reverse order as well,
negating the first reversal. However, this is quite brittle. Thus, I
added a hard-coded path to execute such one-shot callbacks in the
correct order. This system may undergo further changes as we integrate
support for xact callbback API into omni.

For reference, this is how we can generate multiple builds of omni_httpd
for manual testing:

```
cmake /path/to/omnigres -DOMNI_HTTPD_VERSION=v1 && make -j all prepare_omni_httpd
cmake /path/to/omnigres -DOMNI_HTTPD_VERSION=v2 && make -j all prepare_omni_httpd
```

This way you will have two versions of omni_httpd built (v1 and v2), so
that you can now do something like this:

```postgresql
create extension omni_httpd version 'v1';
alter extension omni_httpd update to 'v2';
```

Please note that you still need `create extension omni` to be done prior
to `alter extension` for the upgrade to work. This is a limitation we
should address shortly.